+++ app/src/main/java/mega/privacy/android/app/lollipop/megachat/ChatUploadService.java
@@ -369,15 +369,7 @@ public class ChatUploadService extends Service implements MegaTransferListenerIn
 					Bitmap scaleBitmap = Bitmap.createScaledBitmap(fileBitmap, (int)width, (int)height, false);
 					if (scaleBitmap != null) {
 						log("DATA connection scaled Bitmap != null");
-						File defaultDownloadLocation = null;
-						if (Environment.getExternalStorageDirectory() != null) {
-							defaultDownloadLocation = new File(Environment.getExternalStorageDirectory().getAbsolutePath() + "/" + chatTempDIR + "/");
-						}
-						else{
-							defaultDownloadLocation = getFilesDir();
-						}
-						defaultDownloadLocation.mkdirs();
-						File outFile = new File(defaultDownloadLocation.getAbsolutePath(), file.getName());
+						File outFile = buildChatTempFile(getApplicationContext(), file.getName());
 						if (outFile != null) {
 							log("DATA connection new file != null");
 							FileOutputStream fOut;
@@ -446,22 +438,13 @@ public class ChatUploadService extends Service implements MegaTransferListenerIn
 			try {
 				totalVideos++;
 				numberVideosPending++;
-				File defaultDownloadLocation = null;
-				if (Environment.getExternalStorageDirectory() != null){
-					defaultDownloadLocation = new File(Environment.getExternalStorageDirectory().getAbsolutePath() + "/" + chatTempDIR + "/");
-				}
-				else{
-					defaultDownloadLocation = getFilesDir();
-				}
-
-				defaultDownloadLocation.mkdirs();
-
-				File outFile = new File(defaultDownloadLocation.getAbsolutePath(), file.getName());
+				File chatTempFolder = getCacheFolder(getApplicationContext(), CHAT_TEMPORAL_FOLDER);
+				File outFile = buildChatTempFile(getApplicationContext(), file.getName());
 				int index = 0;
 				if(outFile!=null){
 					while(outFile.exists()){
 						if(index>0){
-							outFile = new File(defaultDownloadLocation.getAbsolutePath(), file.getName());
+							outFile = new File(chatTempFolder.getAbsolutePath(), file.getName());
 						}
 
 						index++;
@@ -476,7 +459,7 @@ public class ChatUploadService extends Service implements MegaTransferListenerIn
 						else
 							fileName=fileName.concat("_"+index+".mp4");
 
-						outFile = new File(defaultDownloadLocation.getAbsolutePath(), fileName);
+						outFile = new File(chatTempFolder.getAbsolutePath(), fileName);
 					}
 				}
 
@@ -570,33 +553,16 @@ public class ChatUploadService extends Service implements MegaTransferListenerIn
 		log("after stopSelf");
 
 		try{
-			String pathSelfie = Environment.getExternalStorageDirectory().getAbsolutePath() +"/"+ temporalPicDIR;
-			File f = new File(pathSelfie);
-			//Delete recursively all files and folder
-			if (f.exists()) {
-				if (f.isDirectory()) {
-					if(f.list().length<=0){
-						f.delete();
-					}
-				}
-			}
+			//Delete recursively all files and folder-??????
+			deleteCacheFolderIfEmpty(getApplicationContext(), TEMPORAL_FOLDER);
 		}
 		catch (Exception e){
 			log("EXCEPTION: pathSelfie not deleted");
 		}
 
 		try{
-			String pathVideoDownsampling = Environment.getExternalStorageDirectory().getAbsolutePath() +"/"+ chatTempDIR;
-			File fVideo = new File(pathVideoDownsampling);
-			//Delete recursively all files and folder
-			if (fVideo.exists()) {
-				if (fVideo.isDirectory()) {
-					if(fVideo.list().length<=0){
-						fVideo.delete();
-					}
-				}
-
-			}
+			//Delete recursively all files and folder-??????
+			deleteCacheFolderIfEmpty(getApplicationContext(), CHAT_TEMPORAL_FOLDER);
 		}
 		catch (Exception e){
 			log("EXCEPTION: pathVideoDownsampling not deleted");
@@ -936,17 +902,8 @@ public class ChatUploadService extends Service implements MegaTransferListenerIn
 
 					ChatUploadService.this.cancel();
 					log("after cancel");
-					String pathSelfie = Environment.getExternalStorageDirectory().getAbsolutePath() + "/" + temporalPicDIR;
-					File f = new File(pathSelfie);
-					//Delete recursively all files and folder
-					if (f.exists()) {
-						if (f.isDirectory()) {
-							if(f.list().length<=0){
-								f.delete();
-							}
-						}
-					}
-					f.delete();
+					//Delete recursively all files and folder-??????
+					deleteCacheFolderIfEmpty(getApplicationContext(), TEMPORAL_FOLDER);
 				}
 				else{
 					if (error.getErrorCode() == MegaError.API_OK) {
@@ -1077,16 +1034,15 @@ public class ChatUploadService extends Service implements MegaTransferListenerIn
 							}
 						}
 					}
-
-					log("IN Finish: "+transfer.getFileName()+" path: "+transfer.getPath());
-					String pathSelfie = Environment.getExternalStorageDirectory().getAbsolutePath() + "/" + temporalPicDIR;
-					if (transfer.getPath() != null) {
-						if (transfer.getPath().startsWith(pathSelfie)) {
+					File tempPic = getCacheFolder(getApplicationContext(), TEMPORAL_FOLDER);
+					log("IN Finish: " + transfer.getFileName() + "path? " + transfer.getPath());
+					if (isFileAvailable(tempPic) && transfer.getPath() != null) {
+						if (transfer.getPath().startsWith(tempPic.getAbsolutePath())) {
 							File f = new File(transfer.getPath());
 							f.delete();
 						}
 					} else {
-						log("transfer.getPath() is NULL");
+						log("transfer.getPath() is NULL or temporal folder unavailable");
 					}
 				}
